requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

pandas==1.4.3
uvicorn==0.18.2
fastapi==0.79.1
scikit-learn==1.1.2
scipy==1.9.0
pydantic==1.9.2
scikit-surprise==1.1.1
python-multipart==0.0.5
azure-storage-file==2.1.0
